# swagger.yaml
openapi: 3.0.3
info:
  title: NewsBreak Scraper API
  description: API for scraping NewsBreak.com with cyclic workflow using RabbitMQ queues and PostgreSQL storage
  version: 1.0.0
  contact:
    name: NewsBreak Scraper Team
    email: scraper@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.newsbreakscraper.com
    description: Production server
...
paths:
  /scraper/start:
    post:
      tags:
        - Scraper
      summary: Start the scraping process
      operationId: startScraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                initial_urls:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: Initial URLs to start scraping from
                  example: ["https://www.newsbreak.com/news", "https://www.newsbreak.com/politics"]
                batch_size:
                  type: integer
                  default: 100
                  description: Number of URLs to process in each batch
                  example: 100
                worker_count:
                  type: integer
                  default: 5
                  description: Number of worker processes to spawn
                  example: 5
              required:
                - initial_urls
      responses:
        '200':
          description: Scraping process started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      job_id:
                        type: string
                        format: uuid
                        example: "bd08612-1a9d-4a6e-8431-1jobid"
                      status:
                        type: string
                        example: "started"
                      worker_count:
                        type: integer
                        example: 5
                      initial_url_count:
                        type: integer
                        example: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scraper/stop:
    post:
      tags:
        - Scraper
      summary: Stop the scraping process
      operationId: stopScraping
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                job_id:
                  type: string
                  format: uuid
                  description: Specific job ID to stop (optional)
                  example: "bd08612-1a9d-4a6e-8431-1jobid"
                force:
                  type: boolean
                  default: false
                  description: Force stop even if workers are busy
                  example: false
      responses:
        '200':
          description: Scraping process stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Scraping process stopped successfully"
                      stopped_workers:
                        type: integer
                        example: 5
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scraper/status:
    get:
      tags:
        - Scraper
      summary: Get scraping process status and statistics
      operationId: getScrapingStatus
      responses:
        '200':
          description: Scraping status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: ['running', 'stopped', 'paused', 'error']
                        example: "running"
                      statistics:
                        type: object
                        properties:
                          urls_processed:
                            type: integer
                            example: 1250
                          data_extracted:
                            type: integer
                            example: 1180
                          errors_count:
                            type: integer
                            example: 70
                          active_workers:
                            type: integer
                            example: 5
                          queue_sizes:
                            type: object
                            properties:
                              scraper_queue:
                                type: integer
                                example: 450
                              newsbreak_urls_queue:
                                type: integer
                                example: 2340
                              newsbreak_data_queue:
                                type: integer
                                example: 890
                              dlx_queue:
                                type: integer
                                example: 15
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /newsbreak/urls:
    get:
      tags:
        - URL Management
      summary: Retrieve collected NewsBreak URLs with pagination
      operationId: getNewsbreakUrls
      responses:
        '200':
          description: NewsBreak URLs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsbreakUrl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - URL Management
      summary: Add URLs to the scraping queue
      operationId: addUrlsToQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                urls:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: List of URLs to add to the scraping queue
                  example: ["https://www.newsbreak.com/news/123", "https://www.newsbreak.com/politics/456"]
              required:
                - urls
      responses:
        '200':
          description: URLs added to queue successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      urls_added:
                        type: integer
                        example: 2
                      duplicates_skipped:
                        type: integer
                        example: 0
                      queue_size:
                        type: integer
                        example: 452
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /newsbreak/urls/deduplicate:
    post:
      tags:
        - URL Management
      summary: Run deduplication process on collected URLs
      operationId: deduplicateUrls
      responses:
        '200':
          description: Deduplication completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      duplicates_removed:
                        type: integer
                        example: 45
                      unique_urls_remaining:
                        type: integer
                        example: 2340
                      processing_time:
                        type: number
                        format: float
                        example: 2.45
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    NewsbreakUrl:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: NewsBreak URL (primary key)
          example: "https://www.newsbreak.com/news/123456"
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request format
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: Error message
                example: "Invalid request parameters"
              error_code:
                type: string
                description: Internal error code for debugging
                example: "BAD_REQUEST_400"
              timestamp:
                type: string
                format: date-time
                description: Timestamp when error occurred
                example: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: Error message
                example: "Resource not found"
              error_code:
                type: string
                description: Internal error code for debugging
                example: "NOT_FOUND_404"
              timestamp:
                type: string
                format: date-time
                description: Timestamp when error occurred
                example: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: Error message
                example: "Internal server error"
              error_code:
                type: string
                description: Internal error code for debugging
                example: "INTERNAL_ERROR_500"
              timestamp:
                type: string
                format: date-time
                description: Timestamp when error occurred
                example: "2024-01-15T10:30:00Z"